requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.1
djangorestframework-simplejwt>=5.2.2
